+++ CHANGELOG.md
@@ -23,6 +23,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Added new tag name filters `:section` and `:-section` `build/snakeskin`
 * Added new props `rootTag` and `rootAttrs` `iBlock`
+* Now the trait uses `aria` attributes `traits/i-visible`
 
 #### :bug: Bug Fix